backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mysql-connector-python==8.2.0
python-dotenv==1.0.0
redis==5.0.1
paho-mqtt==1.6.1
sqlalchemy==2.0.25
alembic==1.13.1
websockets==12.0
